chore: wholesale upgrade to Cypress 13
Also use a recently maintained fork of cypress-plugin-snapshot
