--- cypress.json
@@ -1,15 +1,0 @@
-{
-  "baseUrl": "http://localhost:5000",
-  "ignoreTestFiles": ["**/__snapshots__/*", "**/__image_snapshots__/*"],
-  "env": {
-    "cypress-plugin-snapshots": {
-      "autoCleanUp": true,
-      "imageConfig": {
-        "createDiffImage": true,
-        "resizeDevicePixelRatio": false,
-        "threshold": 0.0001,
-        "thresholdType": "percent"
-      }
-    }
-  }
-}
