+++ cypress/e2e/core/code/snapshots.cy.js
