+++ cypress/e2e/core/contact_footer/behaviour.cy.js
