+++ cypress/e2e/core/contact_footer/parity.cy.js
