+++ cypress/e2e/core/contact_footer/snapshots.cy.js
