+++ cypress/e2e/core/cookie_message/snapshots.cy.js
