+++ cypress/e2e/core/dropdown_menu/behaviour.cy.js
