+++ cypress/e2e/core/dropdown_menu/snapshot.cy.js
