+++ cypress/e2e/core/flash/behaviour.cy.js
