+++ cypress/e2e/core/flash/snapshot.cy.js
