+++ cypress/e2e/core/footer/parity.cy.js
