+++ cypress/e2e/core/footer/snapshots.cy.js
