+++ cypress/e2e/core/meganav/__image_snapshots__/Content  Platform  has the correct content #0.png
