+++ cypress/e2e/core/meganav/__image_snapshots__/Content  Use cases  has the correct content #0.png
