+++ cypress/e2e/core/meganav/screenshots.cy.js
