+++ cypress/e2e/core/meganav/shared.js
