+++ cypress/e2e/core/meganav/snapshots.cy.js
