+++ cypress/e2e/core/showcase/__image_snapshots__/Content  Developers  has the correct content #0.png
