+++ cypress/e2e/core/showcase/__image_snapshots__/Content  Why Ably  has the correct content #0.png
