+++ cypress/e2e/core/showcase/__image_snapshots__/Showcase  looks as per design in a narrow container #0.png
