+++ cypress/e2e/core/showcase/screenshots.cy.js
