+++ cypress/e2e/core/slider/behaviour.cy.js
