+++ cypress/e2e/core/uptime/snapshots.cy.js
