+++ package.json
@@ -27,8 +27,7 @@
     "change-case": "^4.1.1",
     "commander": "^7.2.0",
     "css-loader": "^5.0.0",
-    "cypress": "^6.4.0",
-    "cypress-plugin-snapshots": "^1.4.4",
+    "cypress": "^13.3.1",
     "eslint": "^7.12.1",
     "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-cypress": "^2.11.2",
@@ -65,6 +64,7 @@
     "release": "./scripts/release.sh"
   },
   "dependencies": {
+    "@mrtkrcm/cypress-plugin-snapshots": "https://github.com/mrtkrcm/cypress-plugin-snapshots#v1.13.0",
     "addsearch-js-client": "^0.7.0",
     "array-flat-polyfill": "^1.0.1",
     "dompurify": "^2.2.9",
